AP_PitchController: return floats

+++ libraries/APM_Control/AP_PitchController.h
@@ -15,8 +15,8 @@ public:
     AP_PitchController(const AP_PitchController &other) = delete;
     AP_PitchController &operator=(const AP_PitchController&) = delete;
 
-	int32_t get_rate_out(float desired_rate, float scaler);
-	int32_t get_servo_out(int32_t angle_err, float scaler, bool disable_integrator);
+	float get_rate_out(float desired_rate, float scaler);
+	float get_servo_out(int32_t angle_err, float scaler, bool disable_integrator);
 
 	void reset_I();
 
@@ -56,6 +56,6 @@ private:
 
     AP_Logger::PID_Info _pid_info;
 
-    int32_t _get_rate_out(float desired_rate, float scaler, bool disable_integrator, float aspeed);
+    float _get_rate_out(float desired_rate, float scaler, bool disable_integrator, float aspeed);
     float   _get_coordination_rate_offset(float &aspeed, bool &inverted) const;
 };
